Refactored click views into test source
